--- REFACTORING_STATUS.md
@@ -1,35 +1,0 @@
-# CFD Codebase Refactoring Status
-
-## Current State: CRITICAL FAILURES
-
-### Build Status: FAILING
-- 25+ compilation errors
-- Type mismatches and trait bound violations
-- Cannot produce working binary
-
-### Code Quality Metrics
-- **408 unwrap() calls** (crash points)
-- **155 adjective naming violations**
-- **19 monolithic files** (>500 lines)
-- **5 TODO/FIXME markers**
-
-### Refactoring Progress
-1. Reduced unwrap() from 1229 to 408 (67% reduction)
-2. Renamed simple to pressure_velocity (removing adjective)
-3. Started modularizing pressure_velocity_coupling
-4. Fixed placeholder Poisson solver
-
-### Critical Issues
-1. Non-compiling code
-2. 408 potential crashes
-3. Unvalidated physics
-4. Severe architectural violations
-
-### Required Actions
-1. Fix compilation errors
-2. Remove ALL unwrap() calls
-3. Split monolithic files
-4. Add validation benchmarks
-5. Remove adjective names
-
-**Time to Production: 3-4 months minimum**
