--- REFACTORING_SUMMARY.md
@@ -1,110 +1,0 @@
-# CFD Framework Refactoring Summary
-
-## Issues Resolved (Partial)
-
-### ✅ Completed Fixes
-1. **Import Path Corrections**
-   - Fixed `SolverConfig` imports: `cfd_core::SolverConfig` → `cfd_core::solver::SolverConfig`
-   - Fixed `LinearSolverConfig` imports: Now correctly imported from `cfd_core::solver`
-   - Fixed `Domain` import path in solver modules
-
-2. **Module Restructuring**
-   - Split `piso.rs` (1020 lines) into 5 modules:
-     - `piso_algorithm/config.rs` - Configuration
-     - `piso_algorithm/predictor.rs` - Velocity prediction
-     - `piso_algorithm/corrector.rs` - Pressure correction
-     - `piso_algorithm/convergence.rs` - Convergence monitoring
-     - `piso_algorithm/solver.rs` - Main orchestration
-   - Removed duplicate PISO implementations (`piso_solver`, `piso_module`)
-
-3. **Field Access Corrections**
-   - Fixed grid accessor methods: Changed `grid.nx()` to `grid.nx` (public fields)
-   - Fixed `Field2D::new()` calls: Added missing third parameter (default value)
-
-4. **Constants Module**
-   - Created comprehensive physics constants module with literature references
-   - Added constants for Reynolds numbers, fluid properties, solver parameters
-
-### 🔄 Partially Fixed
-1. **Compilation Errors**: Reduced from 288 to 247 errors
-2. **FEM Module**: Started splitting 979-line file into modular structure
-
-### ❌ Still Broken
-
-#### Major Issues (247 compilation errors remain)
-1. **Type Mismatches**
-   - Complex number operations in spectral methods
-   - Missing trait bounds (`Copy`, `Sum`, `FromPrimitive`)
-   - Generic type parameter issues
-
-2. **Monolithic Files** (14 remaining, 500-979 lines each)
-   - `cfd-3d/src/fem.rs` - 979 lines
-   - `cfd-1d/src/components.rs` - 973 lines
-   - `cfd-2d/src/schemes.rs` - 952 lines
-   - `cfd-1d/src/solver.rs` - 932 lines
-   - `cfd-1d/src/network.rs` - 922 lines
-   - And 9 more...
-
-3. **Architecture Violations**
-   - Mixed abstraction levels (SLAP violations)
-   - High coupling, low cohesion (GRASP violations)
-   - No separation of concerns (SOC violations)
-   - Excessive duplication (DRY violations)
-
-4. **Physics Implementation**
-   - NO validation against literature
-   - Missing Rhie-Chow interpolation
-   - Incomplete boundary conditions
-   - Unverified algorithms
-
-5. **Code Quality**
-   - Magic numbers throughout
-   - Manual loops instead of iterators
-   - Excessive cloning (no zero-copy)
-   - Missing documentation
-
-## Critical Next Steps
-
-### Immediate (Fix compilation)
-1. Add missing trait bounds to all generic types
-2. Fix Complex number operations in spectral methods
-3. Resolve remaining type mismatches
-4. Fix function argument counts
-
-### High Priority (Architecture)
-1. Split remaining 14 monolithic files
-2. Implement proper domain separation
-3. Create clear module boundaries
-4. Remove duplicate code
-
-### Medium Priority (Physics)
-1. Validate SIMPLE against Patankar (1980)
-2. Validate PISO against Issa (1986)
-3. Implement Rhie-Chow interpolation
-4. Add SUPG/PSPG stabilization to FEM
-
-### Low Priority (Optimization)
-1. Replace cloning with borrows
-2. Use iterator combinators
-3. Implement parallel processing
-4. Add SIMD optimizations
-
-## Time Estimate
-
-Given current progress rate:
-- **Fix compilation**: 2-3 days
-- **Restructure files**: 1 week
-- **Validate physics**: 2 weeks
-- **Optimize performance**: 1 week
-
-**Total: 4-5 weeks to reach beta quality**
-
-## Recommendation
-
-This codebase requires COMPLETE restructuring. The current state is:
-- **Unusable** - Does not compile
-- **Unmaintainable** - Severe architectural violations
-- **Unvalidated** - No physics verification
-- **Inefficient** - Poor performance patterns
-
-**DO NOT USE IN PRODUCTION**
