+++ crates/cfd-1d/src/solver/mod.rs
@@ -1,0 +1,44 @@
+//! Network solver for 1D CFD analysis
+
+mod domain;
+mod problem;
+
+pub use domain::NetworkDomain;
+pub use problem::NetworkProblem;
+pub use cfd_core::NetworkSolverConfig as SolverConfig;
+
+use cfd_core::{Result, Solver, Configurable};
+use nalgebra::RealField;
+
+/// Main network solver
+pub struct NetworkSolver<T: RealField> {
+    config: SolverConfig<T>,
+}
+
+impl<T: RealField> NetworkSolver<T> {
+    pub fn new(config: SolverConfig<T>) -> Self {
+        Self { config }
+    }
+}
+
+impl<T: RealField> Solver<T> for NetworkSolver<T> {
+    type Problem = NetworkProblem<T>;
+    type Solution = crate::network::Network<T>;
+    
+    fn solve(&self, _problem: &Self::Problem) -> Result<Self::Solution> {
+        todo!("Implement solver logic")
+    }
+}
+
+impl<T: RealField> Configurable for NetworkSolver<T> {
+    type Config = SolverConfig<T>;
+    
+    fn configure(&mut self, config: Self::Config) -> Result<()> {
+        self.config = config;
+        Ok(())
+    }
+    
+    fn config(&self) -> &Self::Config {
+        &self.config
+    }
+}
