+++ crates/cfd-2d/src/pressure_velocity/rhie_chow.rs
@@ -1,72 +1,132 @@
-//! Rhie-Chow interpolation for pressure-velocity decoupling
+//! Rhie-Chow interpolation for pressure-velocity coupling
 //!
 //! Reference: Rhie, C.M. and Chow, W.L. (1983). "Numerical study of the turbulent
 //! flow past an airfoil with trailing edge separation." AIAA Journal, 21(11), 1525-1532.
 
 use nalgebra::{RealField, Vector2};
+use num_traits::FromPrimitive;
 use crate::grid::StructuredGrid2D;
+use crate::fields::Field2D;
 
-/// Rhie-Chow interpolation for colocated grids
+/// Complete Rhie-Chow interpolation with momentum coefficients
 pub struct RhieChowInterpolation<T: RealField> {
     /// Grid dimensions
     nx: usize,
-    _phantom: std::marker::PhantomData<T>,
     ny: usize,
+    /// Momentum equation coefficients (A_p from discretized momentum equation)
+    ap_coefficients: Field2D<T>,
 }
 
-impl<T: RealField + Copy> RhieChowInterpolation<T> {
-    /// Create new interpolator
+impl<T: RealField + FromPrimitive + Copy> RhieChowInterpolation<T> {
+    /// Create new interpolator with momentum coefficients
     pub fn new(grid: &StructuredGrid2D<T>) -> Self {
         Self {
             nx: grid.nx,
-            _phantom: std::marker::PhantomData,
             ny: grid.ny,
+            ap_coefficients: Field2D::new(grid.nx, grid.ny, T::one()),
         }
     }
     
-    /// Interpolate face velocity with pressure gradient correction
+    /// Update momentum equation coefficients from discretized momentum equation
+    /// A_p is the diagonal coefficient from momentum discretization
+    pub fn update_coefficients(&mut self, ap: &Field2D<T>) {
+        self.ap_coefficients = ap.clone();
+    }
+    
+    /// Compute face velocity with complete Rhie-Chow interpolation
     /// 
-    /// u_f = ū_f - D_f * (∇p)_f
+    /// The complete formulation is:
+    /// u_f = ū_f + d_f * [(∇p)_P - (∇p)_f]
     /// 
     /// where:
     /// - ū_f is the linearly interpolated velocity
-    /// - D_f is the interpolated pressure gradient coefficient
-    /// - (∇p)_f is the pressure gradient at the face
-    pub fn face_velocity(
+    /// - d_f = (Volume/A_p)_f is the interpolated pressure gradient coefficient
+    /// - (∇p)_P is the cell-centered pressure gradient
+    /// - (∇p)_f is the face pressure gradient
+    /// 
+    /// This prevents pressure-velocity decoupling on colocated grids
+    pub fn face_velocity_x(
         &self,
-        u: &Vec<Vec<Vector2<T>>>,
-        p: &Vec<Vec<T>>,
-        d: &Vec<Vec<T>>,
+        u: &Field2D<Vector2<T>>,
+        p: &Field2D<T>,
         dx: T,
+        i: usize,
+        j: usize,
+    ) -> T {
+        // East face between cells (i,j) and (i+1,j)
+        
+        // Linear interpolation of velocity
+        let u_bar = (u.at(i, j).x + u.at(i + 1, j).x) / T::from_f64(2.0).unwrap_or_else(T::zero);
+        
+        // Interpolate pressure gradient coefficient d_f = (Volume/A_p)_f
+        let volume = dx * dx; // 2D cell volume (assuming square cells)
+        let d_p = volume / self.ap_coefficients.at(i, j);
+        let d_e = volume / self.ap_coefficients.at(i + 1, j);
+        let d_face = (d_p + d_e) / T::from_f64(2.0).unwrap_or_else(T::zero);
+        
+        // Cell-centered pressure gradients (from momentum equation)
+        let dp_dx_p = if i > 0 {
+            (p.at(i + 1, j) - p.at(i - 1, j)) / (T::from_f64(2.0).unwrap_or_else(T::zero) * dx)
+        } else {
+            (p.at(i + 1, j) - p.at(i, j)) / dx
+        };
+        
+        let dp_dx_e = if i + 1 < self.nx - 1 {
+            (p.at(i + 2, j) - p.at(i, j)) / (T::from_f64(2.0).unwrap_or_else(T::zero) * dx)
+        } else {
+            (p.at(i + 1, j) - p.at(i, j)) / dx
+        };
+        
+        // Average cell-centered gradient
+        let dp_dx_cells = (dp_dx_p + dp_dx_e) / T::from_f64(2.0).unwrap_or_else(T::zero);
+        
+        // Face pressure gradient
+        let dp_dx_face = (p.at(i + 1, j) - p.at(i, j)) / dx;
+        
+        // Complete Rhie-Chow interpolation
+        u_bar + d_face * (dp_dx_cells - dp_dx_face)
+    }
+    
+    /// Compute face velocity in y-direction
+    pub fn face_velocity_y(
+        &self,
+        v: &Field2D<Vector2<T>>,
+        p: &Field2D<T>,
         dy: T,
         i: usize,
         j: usize,
-        face: Face,
     ) -> T {
-        match face {
-            Face::East => {
-                // East face between cells (i,j) and (i+1,j)
-                let u_bar = (u[i][j].x + u[i+1][j].x) / (T::one() + T::one());
-                let d_face = (d[i][j] + d[i+1][j]) / (T::one() + T::one());
-                let dp_dx = (p[i+1][j] - p[i][j]) / dx;
-                u_bar - d_face * dp_dx
-            }
-            Face::North => {
-                // North face between cells (i,j) and (i,j+1)
-                let v_bar = (u[i][j].y + u[i][j+1].y) / (T::one() + T::one());
-                let d_face = (d[i][j] + d[i][j+1]) / (T::one() + T::one());
-                let dp_dy = (p[i][j+1] - p[i][j]) / dy;
-                v_bar - d_face * dp_dy
-            }
-            _ => T::zero(),
-        }
+        // North face between cells (i,j) and (i,j+1)
+        
+        // Linear interpolation of velocity
+        let v_bar = (v.at(i, j).y + v.at(i, j + 1).y) / T::from_f64(2.0).unwrap_or_else(T::zero);
+        
+        // Interpolate pressure gradient coefficient
+        let volume = dy * dy;
+        let d_p = volume / self.ap_coefficients.at(i, j);
+        let d_n = volume / self.ap_coefficients.at(i, j + 1);
+        let d_face = (d_p + d_n) / T::from_f64(2.0).unwrap_or_else(T::zero);
+        
+        // Cell-centered pressure gradients
+        let dp_dy_p = if j > 0 {
+            (p.at(i, j + 1) - p.at(i, j - 1)) / (T::from_f64(2.0).unwrap_or_else(T::zero) * dy)
+        } else {
+            (p.at(i, j + 1) - p.at(i, j)) / dy
+        };
+        
+        let dp_dy_n = if j + 1 < self.ny - 1 {
+            (p.at(i, j + 2) - p.at(i, j)) / (T::from_f64(2.0).unwrap_or_else(T::zero) * dy)
+        } else {
+            (p.at(i, j + 1) - p.at(i, j)) / dy
+        };
+        
+        // Average cell-centered gradient
+        let dp_dy_cells = (dp_dy_p + dp_dy_n) / T::from_f64(2.0).unwrap_or_else(T::zero);
+        
+        // Face pressure gradient
+        let dp_dy_face = (p.at(i, j + 1) - p.at(i, j)) / dy;
+        
+        // Complete Rhie-Chow interpolation
+        v_bar + d_face * (dp_dy_cells - dp_dy_face)
     }
-}
-
-/// Cell face enumeration
-pub enum Face {
-    East,
-    West,
-    North,
-    South,
 }
