+++ crates/cfd-3d/src/fem/config.rs
@@ -18,6 +18,23 @@ pub struct FemConfig<T: RealField> {
     pub dt: Option<T>,
     /// Reynolds number (for scaling)
     pub reynolds: Option<T>,
+    /// Element type to use
+    pub element_type: ElementType,
+    /// Quadrature order
+    pub quadrature_order: usize,
+}
+
+/// Element types supported
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+pub enum ElementType {
+    /// Linear tetrahedron (4 nodes)
+    Tet4,
+    /// Quadratic tetrahedron (10 nodes)
+    Tet10,
+    /// Linear hexahedron (8 nodes)
+    Hex8,
+    /// Quadratic hexahedron (20 nodes)
+    Hex20,
 }
 
 impl<T: RealField + FromPrimitive> Default for FemConfig<T> {
@@ -28,6 +45,8 @@ impl<T: RealField + FromPrimitive> Default for FemConfig<T> {
             tau: T::from_f64(constants::DEFAULT_STABILIZATION).unwrap_or_else(|| T::zero()),
             dt: None,
             reynolds: None,
+            element_type: ElementType::Tet4,
+            quadrature_order: 2,
         }
     }
 }
